optimize item detection
checking if a player is holding specific items is an expensive operation compared to other conditions. this commit reorders the conditions to fail first on whether the item is active, so only the active and unobtained items (25 at most) have to be searched for, instead of all implemented items. this reduces the size of the per-second lag spikes significantly in-game and will scale better as more items get implemented.