do not clear display items from players, prevent them from scoring points
removing the display items from players is a recurring lag spike that we want to avoid. the problem with this is that players can take the display items to score points. so, we make point scoring only run while the game is in progress. i won't call this a strict optimization, because it does make the change that players are able to possess the display items in the lobby. since tools cannot be used and blocks cannot be placed in adventure mode, this has few consequences - though enderpearls can be thrown, i don't think this matters.
this eliminates lag spikes in the lobby, both from removing the display items and checking for bingo scores. big win for a small compromise?